feature: added handlers on battle

+++ src/network/packets/set-battle-statistics-dm-cc.ts
@@ -46,13 +46,13 @@ export class SetBattleStatisticsDMCCPacket extends Packet {
     public encode() {
         const bytes = new ByteArray();
 
-        bytes.writeShort(this.users.length);
+        bytes.writeInt(this.users.length);
 
         this.users.forEach((user: IUser) => {
             bytes.writeInt(ChatModeratorLevel.LEVELS.indexOf(user.chatModeratorLevel));
             bytes.writeInt(user.deaths);
             bytes.writeInt(user.kills);
-            bytes.writeInt(user.rank);
+            bytes.writeByte(user.rank);
             bytes.writeInt(user.score);
             bytes.writeString(user.name);
         });
